Move stringifyNumber() to src/stringify.js
